refactor: move bad proposal tests 5G and 5H to independent tests
