+++ tests/govtool-frontend/playwright/tests/5-proposal-functionality/proposalFunctionality.dRep.spec.ts
@@ -12,17 +12,13 @@ import GovernanceActionDetailsPage from "@pages/governanceActionDetailsPage";
 import GovernanceActionsPage from "@pages/governanceActionsPage";
 import { Page, expect } from "@playwright/test";
 import kuberService from "@services/kuberService";
-import {
-  BootstrapGovernanceActionType,
-  GovernanceActionType,
-} from "@types";
+import { BootstrapGovernanceActionType, GovernanceActionType } from "@types";
 import walletManager from "lib/walletManager";
-
-const invalidInfinityProposals = require("../../lib/_mock/invalidInfinityProposals.json");
 
 test.beforeEach(async () => {
   await setAllureEpic("5. Proposal functionality");
   await skipIfNotHardFork();
+  test.skip(environments.networkId === 1);
 });
 
 test.describe("Proposal checks", () => {
@@ -125,47 +121,6 @@ test.describe("Proposal checks", () => {
         ).not.toEqual(randomContext);
       }
     });
-  });
-});
-
-test.describe("Bad Proposals", () => {
-  test.use({ storageState: ".auth/dRep01.json", wallet: dRep01Wallet });
-
-  let govActionsPage: GovernanceActionsPage;
-
-  test.beforeEach(async ({ page }) => {
-    await page.route("**/proposal/list?**", async (route) =>
-      route.fulfill({
-        body: JSON.stringify(invalidInfinityProposals),
-      })
-    );
-
-    govActionsPage = new GovernanceActionsPage(page);
-    await govActionsPage.goto();
-  });
-
-  test("5G. Should show warning in bad governance action proposal to the users to visit the site at their own risk, when external url is opened", async () => {
-    const govActionDetailsPage = await govActionsPage.viewFirstProposal();
-
-    await govActionDetailsPage.externalModalBtn.click();
-
-    await expect(govActionDetailsPage.externalLinkModal).toBeVisible();
-    await expect(
-      govActionDetailsPage.currentPage.getByText("Be careful", {
-        exact: false,
-      })
-    ).toBeVisible();
-  });
-
-  test("5H. Should open a new tab in Bad governance action proposal, when external URL is opened", async ({
-    page,
-  }) => {
-    const govActionDetailsPage = await govActionsPage.viewFirstProposal();
-
-    await govActionDetailsPage.externalModalBtn.click();
-    await govActionDetailsPage.continueModalBtn.click();
-    const existingPages = page.context().pages();
-    expect(existingPages).toHaveLength(1);
   });
 });
 
@@ -287,9 +242,9 @@ test.describe("Bootstrap phase", () => {
   test("5L. Should restrict dRep votes to Info Governance actions During Bootstrapping Phase", async ({
     browser,
   }) => {
-    const voteBlacklistOptions = Object.keys(BootstrapGovernanceActionType).filter(
-      (option) => option !== GovernanceActionType.InfoAction
-    );
+    const voteBlacklistOptions = Object.keys(
+      BootstrapGovernanceActionType
+    ).filter((option) => option !== GovernanceActionType.InfoAction);
 
     await Promise.all(
       voteBlacklistOptions.map(async (voteBlacklistOption) => {
